software/requirements.txt
PyQt5
PySerial